a workflow using browserify to bundle js dependencies and keep the code on src organized
