+++ package.json
@@ -26,15 +26,18 @@
   },
   "private": true,
   "devDependencies": {
+    "browserify": "^3.41.0",
     "cordova": "^3.4.0-0.1.3",
     "gulp": "^3.6.0",
     "gulp-assemble": "git://github.com/assemble/gulp-assemble.git",
     "gulp-sass": "^0.7.1",
     "gulp-util": "^2.2.14",
     "handlebars-helpers": "git://github.com/assemble/handlebars-helpers.git#v0.6.0",
+    "jquery": "^2.1.0",
     "lodash": "^2.4.1",
     "mkdirp": "^0.3.5",
-    "q": "^1.0.1"
+    "q": "^1.0.1",
+    "through2": "^0.4.1"
   },
   "jshintConfig": {
     "globals": {
